# pkg/grpc/api/openapi/porterd/v1/porterd.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "porterd-api",
    "description": "porterd",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "PorterService"
    }
  ],
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/execute": {
      "post": {
        "summary": "execute action",
        "description": "perform execute action",
        "operationId": "PorterService_Execute",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1ExecuteResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1ExecuteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ExecuteRequest"
            }
          }
        ],
        "tags": [
          "Installations"
        ]
      }
    },
    "/api/v1/install": {
      "post": {
        "summary": "install action",
        "description": "perform install action",
        "operationId": "PorterService_Install",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1IntallResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1IntallResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1IntallRequest"
            }
          }
        ],
        "tags": [
          "Installations"
        ]
      }
    },
    "/api/v1/uninstall": {
      "post": {
        "summary": "uninstall action",
        "description": "perform uninstall action",
        "operationId": "PorterService_UnInstall",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1UnIntallResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1UnIntallResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UnIntallRequest"
            }
          }
        ],
        "tags": [
          "Installations"
        ]
      }
    },
    "/api/v1/upgrade": {
      "post": {
        "summary": "upgrade action",
        "description": "perform upgrade action",
        "operationId": "PorterService_Upgrade",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1UpgradeResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1UpgradeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpgradeRequest"
            }
          }
        ],
        "tags": [
          "Installations"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ExecuteRequest": {
      "type": "object",
      "properties": {
        "executing": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Execute"
    },
    "v1ExecuteResponse": {
      "type": "object",
      "properties": {
        "output": {
          "type": "string"
        }
      }
    },
    "v1IntallRequest": {
      "type": "object",
      "properties": {
        "installing": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Install"
    },
    "v1IntallResponse": {
      "type": "object",
      "properties": {
        "output": {
          "type": "string"
        }
      }
    },
    "v1UnIntallRequest": {
      "type": "object",
      "properties": {
        "uninstalling": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Uninstall"
    },
    "v1UnIntallResponse": {
      "type": "object",
      "properties": {
        "output": {
          "type": "string"
        }
      }
    },
    "v1UpgradeRequest": {
      "type": "object",
      "properties": {
        "upgrading": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "Upgrade"
    },
    "v1UpgradeResponse": {
      "type": "object",
      "properties": {
        "output": {
          "type": "string"
        }
      }
    }
  }
}